+++ helper/handler.go
@@ -17,11 +17,13 @@ type Handlers struct {
 
 	// ===== CONFIG =====
 	GetActiveConfigByModuleName func(request domain.GetByModuleNameRequest) (*entity.Config, error)             `method:"get_active_config_by_module_name" group:"config" inner:"true"`
+	GetConfigsByModuleId        func(request domain.GetByModuleIdRequest) ([]entity.Config, error)              `method:"get_configs_by_module_id" group:"config" inner:"true"`
 	CreateUpdateConfig          func(config domain.CreateUpdateConfigRequest) (*domain.ConfigModuleInfo, error) `method:"create_update_config" group:"config" inner:"true"`
 	MarkConfigAsActive          func(identity domain.ConfigIdRequest) (*entity.Config, error)                   `method:"mark_config_as_active" group:"config" inner:"true"`
 	DeleteConfig                func(identities []string) (*domain.DeleteResponse, error)                       `method:"delete_config" group:"config" inner:"true"`
 	DeleteVersion               func(identity domain.ConfigIdRequest) (*domain.DeleteResponse, error)           `method:"delete_version" group:"config" inner:"true"`
 	GetAllVersion               func(identity domain.ConfigIdRequest) ([]entity.VersionConfig, error)           `method:"get_all_version" group:"config" inner:"true"`
+	GetConfigByID               func(identity domain.ConfigIdRequest) (entity.Config, error)                    `method:"get_config_by_id" group:"config" inner:"true"`
 
 	// ===== COMMON CONFIG =====
 	GetCommonConfigs         func(identities []string) []entity.CommonConfig                              `method:"get_configs" group:"common_config" inner:"true"`
@@ -45,11 +47,13 @@ func GetHandlers() *Handlers {
 		BroadcastEvent:  controller.Module.BroadcastEvent,
 
 		GetActiveConfigByModuleName: controller.Config.GetActiveConfigByModuleName,
+		GetConfigsByModuleId:        controller.Config.GetConfigsByModuleId,
 		CreateUpdateConfig:          controller.Config.CreateUpdateConfig,
 		MarkConfigAsActive:          controller.Config.MarkConfigAsActive,
 		DeleteConfig:                controller.Config.DeleteConfigs,
 		DeleteVersion:               controller.Config.DeleteConfigVersion,
 		GetAllVersion:               controller.Config.GetAllVersion,
+		GetConfigByID:               controller.Config.GetConfigById,
 
 		GetCommonConfigs:         controller.CommonConfig.GetConfigs,
 		CreateUpdateCommonConfig: controller.CommonConfig.CreateUpdateConfig,